refactor: order 패키지 설계 및 클래스 이동
