+++ src/main/java/kitchenpos/dao/TableGroupDao.java
@@ -1,6 +1,6 @@
 package kitchenpos.dao;
 
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.order.TableGroup;
 
 import java.util.List;
 import java.util.Optional;
