+++ src/main/java/kitchenpos/domain/order/TableGroup.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
+
+import kitchenpos.domain.order.OrderTable;
 
 import java.time.LocalDateTime;
 import java.util.List;
